Allow for 32 bit trace source input

+++ ext/ddtrace.h
@@ -148,7 +148,7 @@ ZEND_BEGIN_MODULE_GLOBALS(ddtrace)
     HashTable telemetry_spans_created_per_integration;
     ddog_SidecarActionsBuffer *telemetry_buffer;
 
-    uint8_t trace_source_bm;
+    uint32_t trace_source_bm;
     bool asm_event_emitted;
 
 #if PHP_VERSION_ID >= 80000
